refactor: move clients components into own files
