+++ src/lib/modules/providers/components/Clients/ActionConfirmationModal/index.ts
@@ -1,0 +1,1 @@
+export * from './ActionConfirmationModal';
