+++ src/lib/modules/providers/components/Clients/ClientList/index.ts
@@ -1,0 +1,1 @@
+export * from './ClientList';
