+++ src/lib/modules/providers/components/Clients/ClientList/ui/index.ts
@@ -1,0 +1,2 @@
+export * from './ListCells';
+export * from './ClientListItem';
