+++ src/lib/modules/providers/components/Clients/MemberDetailsModal/index.ts
@@ -1,0 +1,1 @@
+export * from './MemberDetailsModal';
