+++ src/lib/modules/providers/components/Clients/index.ts
@@ -1,0 +1,3 @@
+export * from './ActionConfirmationModal';
+export * from './ClientList';
+export * from './MemberDetailsModal';
